Check if entity exists before delete action

+++ backend/src/modules/customers/services/DeleteCustomerService.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from 'tsyringe';
 
+import AppError from '../../../shared/errors/AppError';
 import ICustomersRepository from '../repositories/ICustomersRepository';
 
 @injectable()
@@ -12,6 +13,14 @@ class DeleteCustomerService {
 
   // Serviço para remover um clientes pelo seu ID
   public async execute(id: string): Promise<string> {
+    // Verificando se o cliente existe antes de apagar
+    const customerToDelete = await this.customersRepository.findById(id);
+
+    if (!customerToDelete) {
+      throw new AppError('Customer not found.', 404);
+    }
+
+    // Apagando o cliente
     const responseMessage = await this.customersRepository.delete(id);
 
     return responseMessage;
